fix(runtime): apply all critical fixes from runtime security review
Apply 55 critical fixes from docs/runtime-review.md across three
categories. All 1182 tests pass.

Security fixes (S-01–S-23):
- S-11: ReDoS — regex step limit (RE_MAX_STEPS=1,000,000)
- S-13: XSS — block javascript:/data:/vbscript: URLs in Markdown
- S-14: TOML rt_toml_is_valid returns 0 for invalid input
- S-15: TOML memcpy type-punning fix preserving get/get_str
- S-16: JSON recursion depth limit (200 levels, soft-error via flag)
- S-17: XML element nesting depth limit (200 levels)
- S-18: YAML recursion depth limit
- S-20: Decompression bomb — output size cap
- S-01–S-10, S-12, S-19, S-21–S-23: null guards, integer overflow
  checks, OOB guards, and safe string handling across crypto, TLS,
  network, exec, file, and graphics APIs

Reliability fixes (R-01–R-28):
- Null pointer guards and bounds checks across bytes, seq, string,
  map, and numeric runtime APIs
- Proper error propagation for file I/O, network, and crypto
- Integer overflow guards in bigint, mat4, and numeric conversion
- Resource leak fixes in async, scheduler, and concurrency APIs

Optimization correctness fixes (O-01–O-04):
- O-01/O-02: chunked file I/O path for large binary files
- O-03: fmt buffer growth fix
- O-04: rt_xml_text_content O(n) string builder (was O(n²))

XML ownership model (pervasive correctness fix):
- Child nodes were freed immediately after rt_seq_push because
  rt_seq_push does not retain elements. Fixed xml_node_finalizer
  to iterate and release owned children, removed premature releases
  in parse_element/parse_document/rt_xml_set_text, and removed
  incorrect rt_obj_release_check0 calls on borrowed references in
  all iteration, navigation, and serialization functions.
- rt_xml_remove_at: removed double-release (seq_get + seq_remove).

Add ctests: RTSecurityFixesTests.c (59 tests), RTOptimizationFixesTests.c
(26 tests), plus RTArithmeticUBTests, RTBufferOverflowTests,
RTConcurrencyTests, RTGraphicsMemTests, RTNullCheckTests,
RTReallocMemTests stubs registered in CMakeLists.txt.

+++ src/runtime/rt_duration.c
@@ -156,12 +156,12 @@ int64_t rt_duration_div(int64_t duration, int64_t divisor)
 
 int64_t rt_duration_abs(int64_t duration)
 {
-    return duration >= 0 ? duration : -duration;
+    return duration >= 0 ? duration : (int64_t)(-(uint64_t)duration);
 }
 
 int64_t rt_duration_neg(int64_t duration)
 {
-    return -duration;
+    return (int64_t)(-(uint64_t)duration);
 }
 
 //=============================================================================
